Fold call-site wrapping into MigrateItemWriterWrite
Per PR review (single-recipe is preferable), merge MigrateItemWriterWriteCalls
into MigrateItemWriterWrite:

- Widen the precondition to fire on files that either declare or invoke
  ItemWriter.write(List).
- Add a visitMethodInvocation override that wraps List arguments in
  new Chunk<>(...). Because super.visitMethodDeclaration walks children before
  the body reparse, the inner call's methodType is still intact and a plain
  MethodMatcher works — no simpleName workaround needed.
- Skip the parameter-passthrough case (argument is the parameter of the
  enclosing ItemWriter.write(List) declaration) so the migrated Chunk is
  passed through unwrapped instead of round-tripping via .getItems().
- Drop MigrateItemWriterWriteCalls and its YAML chain entry; move its tests
  into MigrateItemWriterWriteTest, renaming the previously-chained ones.
- Regenerate recipes.csv.
