+++ src/main/resources/META-INF/rewrite/spring-batch-5.0.yml
@@ -35,7 +35,6 @@ recipeList:
   - org.openrewrite.java.spring.batch.MigrateStepBuilderFactory
   - org.openrewrite.java.spring.batch.AddTransactionManagerToTaskletAndChunk
   - org.openrewrite.java.spring.batch.MigrateItemWriterWrite
-  - org.openrewrite.java.spring.batch.MigrateItemWriterWriteCalls
   - org.openrewrite.java.spring.batch.RemoveDefaultBatchConfigurer
   - org.openrewrite.java.spring.batch.UpgradeSkipPolicyParameterType
   - org.openrewrite.java.ChangeType:
